autoimprover: simplify winpeas checks
