+++ winPEAS/winPEASexe/winPEAS/Helpers/PermissionsHelper.cs
@@ -25,6 +25,65 @@ namespace winPEAS.Helpers
     /// Get interesting permissions from Files, Folders and Registry
     internal static class PermissionsHelper
     {
+        public static HashSet<string> GetUnprivilegedTokenSids()
+        {
+            bool isPrivilegedContext;
+            return GetUnprivilegedTokenSids(out isPrivilegedContext);
+        }
+
+        public static HashSet<string> GetUnprivilegedTokenSids(out bool isPrivilegedContext)
+        {
+            var unprivilegedSids = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
+            {
+                "S-1-1-0",      // Everyone
+                "S-1-5-4",      // Interactive
+                "S-1-5-11",     // Authenticated Users
+                "S-1-5-32-545", // BUILTIN\Users
+            };
+
+            isPrivilegedContext = true;
+            try
+            {
+                using (WindowsIdentity identity = WindowsIdentity.GetCurrent())
+                {
+                    var principal = new WindowsPrincipal(identity);
+                    var administratorsSid = new SecurityIdentifier(WellKnownSidType.BuiltinAdministratorsSid, null);
+                    isPrivilegedContext = principal.IsInRole(administratorsSid) ||
+                        (identity.User != null && IsPrivilegedServiceSid(identity.User.Value));
+                    if (!isPrivilegedContext && identity.User != null)
+                    {
+                        unprivilegedSids.Add(identity.User.Value);
+                    }
+
+                    if (identity.Groups != null)
+                    {
+                        foreach (IdentityReference group in identity.Groups)
+                        {
+                            var sid = group as SecurityIdentifier;
+                            if (sid != null && principal.IsInRole(sid) && !sid.Equals(administratorsSid) &&
+                                !IsPrivilegedServiceSid(sid.Value))
+                            {
+                                unprivilegedSids.Add(sid.Value);
+                            }
+                        }
+                    }
+                }
+            }
+            catch
+            {
+                // Well-known low-privilege principals still provide useful configuration auditing.
+            }
+
+            return unprivilegedSids;
+        }
+
+        private static bool IsPrivilegedServiceSid(string sid)
+        {
+            return string.Equals(sid, "S-1-5-18", StringComparison.OrdinalIgnoreCase) ||
+                string.Equals(sid, "S-1-5-19", StringComparison.OrdinalIgnoreCase) ||
+                string.Equals(sid, "S-1-5-20", StringComparison.OrdinalIgnoreCase);
+        }
+
         public static List<string> GetPermissionsFile(string path, Dictionary<string, string> SIDs, PermissionType permissionType = PermissionType.DEFAULT)
         {
             /*Permisos especiales para carpetas
